Support blocking handler fn's
If a handler function does heavy computation, or it has blocking IO
in it, it should not block the whole task. Tokio provides the
spawn_blocking and block_in_place functions to support
running blocking operations on a threadpool that does not impact the
async executor. Because the operation will take place on a different
thread, we need to clone out of the buffer that the client uses, so
there is an allocation penalty to be paid for using a blocking
operation, but this is less important than keeping the executor moving.